date, setDate put into context Co-authored by: j2macwilliams Co-authored by: Kennethaustin
