+++ src/contexts/Context.js
@@ -1,16 +1,27 @@
 import React, { useState } from 'react'
+import dayjs from 'dayjs';
+
 
 const StateContext = React.createContext()
 
 const StateProvider = ({ children })  =>  {
   const [selected, setSelected] = useState([])
-
+  const [date, setDate] = useState(dayjs())
+  
+  const currentYear = date.year();
+  const currentMonth = date.month(); // January = 0
 
   return(
     <StateContext.Provider
     value = {{
       selected: selected,
-      setSelected: setSelected
+      setSelected: setSelected,
+
+      date:date,
+      setDate:setDate,
+
+      currentYear:currentYear,
+      currentMonth:currentMonth
     }}
     >
       {children}
